refactor: split momentum config and position modules
